CFD Mesh GUI Makeover and wake modeling updates.
The CFD Mesh GUI had grown complicated and too large for laptop & projector
screens.  This makeover to a tab-based format shoudl simplify the gui and
make it usable on low-resolution screens.

Also, the wake modeling has been updated and improved.

+++ src/vsp/Mesh.cpp
@@ -71,7 +71,7 @@ void Mesh::Clear()
 
 void Mesh::LimitTargetEdgeLength( Node* n )
 {
-	for( int i = 0; i < n->edgeVec.size(); i++ )
+	for( int i = 0; i < (int)n->edgeVec.size(); i++ )
 	{
 		LimitTargetEdgeLength( n->edgeVec[i], n );
 	}
@@ -1129,7 +1129,7 @@ void Mesh::AdjustEdgeLengths()
 
 void Mesh::ComputeTargetEdgeLength( Node* n )
 {
-	for( int i = 0; i < n->edgeVec.size(); i++)
+	for( int i = 0; i < (int)n->edgeVec.size(); i++)
 	{
 		ComputeTargetEdgeLength( n->edgeVec[i] );
 	}
